+++ engines/hopkins/lines.h
@@ -85,7 +85,7 @@ public:
 	int CONTOURNE(int a1, int a2, int a3, int a4, int a5, int16 *a6, int a7);
 	int MIRACLE(int a1, int a2, int a3, int a4, int a5);
 	int GENIAL(int a1, int a2, int a3, int a4, int a5, int a6, int a7, int16 *a8, int a9);
-	byte *PARCOURS2(int a1, int a2, int a3, int a4);
+	int16 *PARCOURS2(int a1, int a2, int a3, int a4);
 	int PARC_PERS(int a1, int a2, int a3, int a4, int a5, int a6, int a7);
 	int VERIF_SMOOTH(int a1, int a2, int a3, int a4);
 	int SMOOTH_MOVE(int a1, int a2, int a3, int a4, int a5, int a6);
